Modificacion visualizacion.h cambio de variables

+++ visualizacion.h
@@ -1,7 +1,8 @@
+#include <stdint.h>
 /**
 *\ brief Funcion para visualizar los registros
 *\ param *R Puntero que señala la posicion de los registros almacenados.
 *\ param  *Banderas Corresponde a las banderas Cero, Negativo, Acarreo y Sobreflujo.
 *\ return void 
 */
-void visualizacion_registro(unsigned long *R, unsigned long *Banderas);
+void visualizacion_registro(uint32_t *R, uint32_t *Banderas);
